Extract DdTrace from foundation
